refactor(steps-11-15): decompose service.py + router handler extraction
Service decomposition (Steps 11-15): 978-line service.py monolith →
service/ package with 5 classes:
- ProcessManager (process.py): pid management, port detection
- KeysEnvWriter (keys_env.py): secure keys.env writing
- LaunchdBackend (launchd.py): macOS plist management
- SystemdBackend (systemd.py): Linux unit management
- ServiceInstaller (installer.py): composition root

Router handler extraction (parallel agent): 692-line router.py → 127
lines + 3 handler modules:
- speech_handlers.py: synthesize, record, voices
- music_handlers.py: music_on/off/play/list/vibe/next
- system_handlers.py: health, chime

router.py: max_complexity 13→9, module_size 692→127.
service.py deleted. test_service.py split into 5 test files.
Tests: 84 service tests (was 80), handler tests preserved.
