+++ .oo-baseline.json
@@ -363,15 +363,80 @@
     "public_attr_violations": 0.0,
     "future_annotations": 1.0
   },
-  "src/punt_vox/service.py": {
+  "src/punt_vox/service/__init__.py": {
     "method_ratio": 0.0,
     "encapsulation_ratio": 1.0,
-    "avg_params": 0.37,
-    "max_complexity": 20.0,
-    "avg_complexity": 3.67,
-    "module_size": 978.0,
+    "avg_params": 0.39,
+    "max_complexity": 1.0,
+    "avg_complexity": 1.0,
+    "module_size": 136.0,
     "classes_per_module": 0.0,
     "class_to_func_ratio": 0.0,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/service/installer.py": {
+    "method_ratio": 0.889,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 0.22,
+    "max_complexity": 3.0,
+    "avg_complexity": 1.89,
+    "module_size": 169.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 0.5,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/service/keys_env.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 1.0,
+    "max_complexity": 20.0,
+    "avg_complexity": 10.5,
+    "module_size": 115.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 1.0,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/service/launchd.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 0.57,
+    "max_complexity": 3.0,
+    "avg_complexity": 1.86,
+    "module_size": 169.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 1.0,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/service/process.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 0.33,
+    "max_complexity": 9.0,
+    "avg_complexity": 4.33,
+    "module_size": 187.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 1.0,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/service/systemd.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 0.6,
+    "max_complexity": 7.0,
+    "avg_complexity": 2.6,
+    "module_size": 236.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
     "public_attr_violations": 0.0,
     "future_annotations": 1.0
@@ -434,9 +499,22 @@
     "avg_params": 0.0,
     "max_complexity": 0.0,
     "avg_complexity": 0.0,
-    "module_size": 96.0,
+    "module_size": 102.0,
     "classes_per_module": 0.0,
     "class_to_func_ratio": 1.0,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/voxd/_parse.py": {
+    "method_ratio": 0.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 2.0,
+    "max_complexity": 2.0,
+    "avg_complexity": 2.0,
+    "module_size": 23.0,
+    "classes_per_module": 0.0,
+    "class_to_func_ratio": 0.0,
     "init_violations": 0.0,
     "public_attr_violations": 0.0,
     "future_annotations": 1.0
@@ -473,7 +551,7 @@
     "avg_params": 1.64,
     "max_complexity": 6.0,
     "avg_complexity": 1.79,
-    "module_size": 281.0,
+    "module_size": 292.0,
     "classes_per_module": 1.0,
     "class_to_func_ratio": 0.5,
     "init_violations": 0.0,
@@ -500,6 +578,19 @@
     "max_complexity": 2.0,
     "avg_complexity": 1.14,
     "module_size": 93.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 1.0,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/voxd/music_handlers.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 2.0,
+    "max_complexity": 10.0,
+    "avg_complexity": 3.57,
+    "module_size": 257.0,
     "classes_per_module": 1.0,
     "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
@@ -535,10 +626,23 @@
   "src/punt_vox/voxd/router.py": {
     "method_ratio": 1.0,
     "encapsulation_ratio": 1.0,
-    "avg_params": 2.05,
+    "avg_params": 1.2,
+    "max_complexity": 9.0,
+    "avg_complexity": 2.8,
+    "module_size": 127.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 1.0,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/voxd/speech_handlers.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 2.2,
     "max_complexity": 13.0,
-    "avg_complexity": 3.48,
-    "module_size": 692.0,
+    "avg_complexity": 4.6,
+    "module_size": 266.0,
     "classes_per_module": 1.0,
     "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
@@ -554,6 +658,19 @@
     "module_size": 407.0,
     "classes_per_module": 1.0,
     "class_to_func_ratio": 0.333,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/voxd/system_handlers.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 2.5,
+    "max_complexity": 3.0,
+    "avg_complexity": 2.0,
+    "module_size": 102.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
     "public_attr_violations": 0.0,
     "future_annotations": 1.0
