+++ tests/test_voxd_system_handlers.py
@@ -1,0 +1,7 @@
+"""Tests for punt_vox.voxd.system_handlers -- chime, voices, health handlers.
+
+NOTE: No existing tests for chime/voices/health handlers were present in the
+original test_voxd_router.py. This file is a placeholder documenting the gap.
+"""
+
+from __future__ import annotations
